requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
opencv-python==4.8.0.76
pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0

# Graph Database Integration
neo4j==5.11.0
networkx==3.1
redis==4.6.0

# Advanced Computer Vision
ultralytics==8.0.196
segmentation-models-pytorch==0.3.2
albumentations==1.3.1
face-recognition==1.3.0
mediapipe==0.10.3
scikit-image==0.21.0

# Stream Processing
kafka-python==2.0.2
aiokafka==0.8.11
prometheus-client==0.17.1
aiohttp==3.8.5

# Container Orchestration
kubernetes==27.2.0
pyyaml==6.0.1

# Data Processing
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.1

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1

# Development
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2